pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "peek"
dynamic = ["version"]
description = "Python Base Libraries - A collection of utilities for Python development"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "kaydxh", email = "kaydxh@example.com"}
]
keywords = [
    "utilities",
    "webserver",
    "grpc",
    "http",
    "networking",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "PyYAML>=6.0,<7.0",
    "click>=8.0.0,<9.0.0",
]

[project.optional-dependencies]
http = [
    "requests>=2.28.0,<3.0.0",
    "httpx>=0.24.0,<1.0.0",
]
web = [
    "fastapi>=0.100.0,<1.0.0",
    "uvicorn[standard]>=0.23.0,<1.0.0",
    "gunicorn>=21.0.0,<24.0.0",
]
grpc = [
    "grpcio>=1.57.0,<2.0.0",
    "grpcio-tools>=1.57.0,<2.0.0",
    "grpcio-health-checking>=1.57.0,<2.0.0",
    "grpcio-reflection>=1.57.0,<2.0.0",
]
otel = [
    "opentelemetry-api>=1.20.0,<2.0.0",
    "opentelemetry-sdk>=1.20.0,<2.0.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-exporter-otlp>=1.20.0,<2.0.0",
    "opentelemetry-semantic-conventions>=0.41b0",
]
database = [
    "sqlalchemy[asyncio]>=2.0.0,<3.0.0",
    "aiomysql>=0.2.0,<1.0.0",
    "redis>=5.0.0,<6.0.0",
]
monitor = [
    "psutil>=5.9.0",
    "pynvml>=11.5.0",
    "matplotlib>=3.7.0",
]
cv = [
    "ffmpeg-python>=0.2.0",
    "av>=10.0.0",
    "opencv-python>=4.8.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "autoflake>=2.0.0",
    "mypy>=1.5.0",
    "flake8>=6.1.0",
    "pylint>=2.17.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
]
prod = [
    "peek[http]",
    "peek[web]",
    "peek[grpc]",
    "peek[otel]",
    "peek[database]",
    "psutil>=5.9.0",
]
all = [
    "peek[prod]",
    "peek[dev]",
    "peek[monitor]",
    "peek[cv]",
]

[project.urls]
Homepage = "https://github.com/kaydxh/peek"
Documentation = "https://github.com/kaydxh/peek#readme"
Repository = "https://github.com/kaydxh/peek.git"
Issues = "https://github.com/kaydxh/peek/issues"

# [project.scripts]
# peek = "peek.cli:main"  # 如果需要 CLI 入口

[tool.setuptools]
package-dir = {"" = "src"}
package-data = {"peek" = ["py.typed"]}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "peek.__version__.__version__"}

# Black 配置
[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

# isort 配置
[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true
known_first_party = ["peek"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

# pytest 配置
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"

# mypy 配置
[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false
check_untyped_defs = false
no_implicit_optional = false
strict = false
exclude = [
    "build",
    "dist",
    "examples",
    "tests",
]

# coverage 配置
[tool.coverage.run]
source = ["src/peek"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
